+++ source/CMakeROOT_6_16.cmake
@@ -87,6 +87,7 @@ add_library(ostap SHARED src/format.cpp
                          src/MatrixUtils.cpp                         
                          src/Models.cpp
                          src/Models2D.cpp
+                         src/Moments.cpp
                          src/MoreMath.cpp
                          src/MoreRooFit.cpp
                          src/MoreRooFit2.cpp
@@ -94,6 +95,7 @@ add_library(ostap SHARED src/format.cpp
                          src/Mute.cpp
                          src/NStatEntity.cpp
                          src/Notifier.cpp
+                         src/Ostap.cpp
                          src/OstapDataFrame.cpp
                          src/P2Quantile.cpp
                          src/Parameterization.cpp