+++ source/src/dict/Ostap.hh
@@ -60,6 +60,7 @@
 #include "Ostap/MatrixTransforms.h"
 #include "Ostap/Models.h"
 #include "Ostap/Models2D.h"
+#include "Ostap/Moments.h"
 #include "Ostap/MoreMath.h"
 #include "Ostap/MoreRooFit.h"
 #include "Ostap/MoreRooFit2.h"